smart_bi/requirements.txt
streamlit==1.38.0
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
plotly==5.24.1
matplotlib==3.9.2
